Add Window class to gfx module
The graphics manager is currenty handling both rendering and window
related tasks. To keep the manager only responsable for rendering and
to simply the class, this commit adds a new Window class. This class
works as a wrapper of the GLFWwindow struct and, a handler of all the
window related functions.