Add Glfw_manager class to gfx module
The graphics_manager is currently having to interface directly with glfw,
to forward calls and initialize/terminate the libary. The issue with this
current state, is that the manager is doing work that is not directly
related to rendering and, having to include the library in it's code,
creating an unwanted dependency between the graphics manager and
the GLFW library.

This commit adds the class Glfw_manager, to work with the Window
class, as an abstraction layer between the Glfw library and the rest of
the engine, that way, we can move the responsabilty of managing the
library, from the graphics manager to this new wrapper class.