Refactor the engine initialization function
Since we added new classes to incapsulate the GLFW library
initialization and, to manage Glfw's windows , the graphics manager
needed to be refactored to accommodate those changes. This refactor
introduced some changes on the initialization code. Therefore, the
engine initialization function had to be refactored to use the
Gflw_manager and the refactored graphics manager.