Refactor gom's CMakeLists to use SID preprocessor
This commit refactors the gom module CMakeLists file, to use the SID preprocessor utility for an aditional build step. With this refactor, the module's source files are preprocessed to generate the final source files to build the engine. Those source files have the SID macros substituted with the corresponding hashed string.