Refactor player gameplay states to use SID utility
This commit refactors the player gameplay states, to use the version of the animator controller functions that have unsigned integer types as parameters, to do this, we are using the new SID preprocessor utility, to generate on a PRE BUILD step, the hashed string values representing the animator controller variables. Previously, we were using the version os these animator controller functions, that receive string values as parameters, so the hashing of those string values were done in real time by the animator controller functions. Now, since those values are computed on a PRE BUILD step, we have significantly improved the perfomance of the player's update function.