Delete Game_coll_listener class
This commit deletes the Game_coll_listener class. This collision listener class, was implemented before we had the SID preprocessor and before the implementation of the Event system, now, with those features, we could implement a much better collision listener class, Engine_collision_listener, to take full advantage of those features, and implement a much better solution to the problem of having game objects react differently, to collisions with different types of game object.