Implement register_game_object member function
This commit implements the game_object_manager register_game_object member function, to be used by the Game_object constructor to get it's handle index and guid. With the Game_objects being responsible to call this member function to setup their data members, the Game_object_manager instantiate member function needed to be refactor to accommodate this changes.