Implement destroy member function
This commit implements the destroy member function on the Game_object class. Now that the Game_object's destructor is a private member function, the destroy member function is the only way to destroy a game_object, calling the Game_object_manager's request_destruction, that guarantees that the object will be destroyed after the current game loop.