Refactor Creator
This commit refactors the Creator class, to acommodate the changes made in the Game_object class. The main change is that the Creator no longer needs to set the Game_object's guid and handle index and, the allocation of memory for the Game_object should be done in the Create member function.