Implement Window's viewport scaling
This commit, implements the Window::update_viewport function, to scale the window's viewport whenever the user change it's size. This scaling function preserves the viewport's aspect ratio, putting black bars whenever the viewport is not able to fully scale to the size of the window without loosing it's aspect ratio.