Refactor to get the entire comma delimited string
Now, the parser will assume that all the characters on the name string should be used as the sprite name, this includes dot characters. Before this change, we were removing any part of the string that followed a dot character, but now, to be able to accept a larger variety of characters, we are removing this restriction.