Add request_level member function
This commit, adds a new member function to the level_manager's interface. Now, when any system wants to load a new level, it should call the request_level function, which basically checks if this new level can be loaded and, if so, sets a flag that will cause the new level to be loaded at the start of the next tick loop.