Rename and refactor the restart function
Since the actual restarting of  a level needs to be done at a specific time, we moved the restart member function to the implementation of the Level_manager. This way, after a request to restart the level, the manager will call this restart function when possible to do so. We also renamed the function to restart_level.