Add flag for invincibility_mode
After the player takes a hit, we want to allow it to have a small time window in which it can not take any damage. The name of this "state" is the invicibility_mode, and to know when the player is in this state, we have added the is_in_invincibility_mode flag data member.