Refactor create function to use a tile map Object
The create function was very limited, with only one parameter for the object's position. This limitation was becoming a problem, mainly because we would like to have a better control of the properties of the object being instantiated by the Creators.

To solve this issue we have decided to use an Object that we already had, but needed just a few adjustments. The Tile_map Objects not only store the basic information regarding an Object on the map, such as position and size, but ALSO stores an unordered_set of custom properties. So by using these Objects, we can have ANY custom property assigned to an instance of an object on the map, done by using the level editor, accessible to the Creators, allowing us to completely customize our Game_objects.