Refactor instantiate function to use Tile_map Object
Because the Creator's create function was refactored to use a Tile_map Object, and the Game_object_manager interfaces directly with the creator, it's instantiate function also needed to be refactored to accommodate this change. Therefore, the instantiate function now also receives a Tile_map Object as its parameter, containing the properties of the objected to be instantantiated. This Object is then passed to the respective creator, which will be responsible to actually create the Game_object.