Refactor to use the new instantiate function
Now that all the necessary data to instantiate the Game_objects on a particular level is stored on Tile_map objects, the level_manager needed to be refactored to use these map Objects to instantiate all the Game_objects on the current level. Therefore, we removed the load_objects_data, since this data is already on the map, and refactored the instantiate_level_objects function to use the Tile_map Objects. Furthermore, since all the necessary data to instantiate those Game_objects is on the Tile_map, the m_level_data data member is no longer necessary.