Copy Engine resources to executable folder
Since the Engine uses it's default resources on initialization, theses files need to be available at runtime when executing the game. Therefore, the CMakeLists file was refactored to copy the Engine resources to the executable folder after building. This way, both the resources for the game and engine are available on the same folder at runtime.