Refactor to only load the Engine's default resources
The Level_manager was being responsible to load not only the Engine's specific resources, default shaders, but also the Game's specific resources, sprite atlases. This created a permanent coupling between the Engine and the Game, that would never allow the Engine to be generic. Therefore, the Level_manager was refactored to only be responsible to load the Engine's resources, in this case the default shaders. The game specific resources are now loaded on the Client code at startup.